refactor: rename MercuryRoomList to YGOProRoomList and simplify SQLite setup
