+++ src/mercury/room/domain/YGOProRoom.ts
@@ -683,8 +683,8 @@ export class YGOProRoom extends YgoRoom {
       draw_count: this._hostInfo.draw_count,
       time_limit: this._hostInfo.time_limit,
       rule: this._hostInfo.rule,
-      no_check: this._hostInfo.no_check_deck,
-      no_shuffle: this._hostInfo.no_shuffle_deck,
+      no_check: Boolean(this._hostInfo.no_check_deck),
+      no_shuffle: Boolean(this._hostInfo.no_shuffle_deck),
       banlist_hash: this.edoBanListHash ?? this.banListHash,
       istart: this.isStart,
       main_min: 40,
